app/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-common-helpers==0.9.2
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
Markdown==3.2.2
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.9
